Added / reordered tests for PATCH method
