+++ src/WireMock.Net/RequestBuilders/IMethodRequestBuilder.cs
@@ -8,12 +8,28 @@ namespace WireMock.RequestBuilders
     public interface IMethodRequestBuilder : IHeadersAndCookiesRequestBuilder
     {
         /// <summary>
+        /// The using delete.
+        /// </summary>
+        /// <returns>
+        /// The <see cref="IRequestBuilder"/>.
+        /// </returns>
+        IRequestBuilder UsingDelete();
+
+        /// <summary>
         /// The using get.
         /// </summary>
         /// <returns>
         /// The <see cref="IRequestBuilder"/>.
         /// </returns>
         IRequestBuilder UsingGet();
+
+        /// <summary>
+        /// The using head.
+        /// </summary>
+        /// <returns>
+        /// The <see cref="IRequestBuilder"/>.
+        /// </returns>
+        IRequestBuilder UsingHead();
 
         /// <summary>
         /// The using post.
@@ -24,12 +40,12 @@ namespace WireMock.RequestBuilders
         IRequestBuilder UsingPost();
 
         /// <summary>
-        /// The using delete.
+        /// The using patch.
         /// </summary>
         /// <returns>
         /// The <see cref="IRequestBuilder"/>.
         /// </returns>
-        IRequestBuilder UsingDelete();
+        IRequestBuilder UsingPatch();
 
         /// <summary>
         /// The using put.
@@ -38,14 +54,6 @@ namespace WireMock.RequestBuilders
         /// The <see cref="IRequestBuilder"/>.
         /// </returns>
         IRequestBuilder UsingPut();
-
-        /// <summary>
-        /// The using head.
-        /// </summary>
-        /// <returns>
-        /// The <see cref="IRequestBuilder"/>.
-        /// </returns>
-        IRequestBuilder UsingHead();
 
         /// <summary>
         /// The using any verb.
